+++ ElvUI_MerathilisUI/Core/Changelog/6.56.lua
@@ -1,6 +1,6 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
-MER.Changelog[653] = {
+MER.Changelog[656] = {
 	RELEASE_DATE = "TBD",
 	FIXES = {},
 	NEW = {},