+++ ElvUI_MerathilisUI/Core/Changelog/Previous/6.53.lua
@@ -1,0 +1,13 @@
+local MER = unpack(ElvUI_MerathilisUI)
+
+MER.Changelog[653] = {
+	RELEASE_DATE = "26.02.2025",
+	FIXES = {
+		"[11.1]: Changes because of the new Patch",
+		"[World Map]: Disable Event Tracker for now, Blizzard got their own one",
+	},
+	NEW = {},
+	IMPROVEMENTS = {
+		"[AutoButtons]: Add more openable items",
+	},
+}