+++ ElvUI_MerathilisUI/Core/Changelog/Previous/6.54.lua
@@ -1,0 +1,10 @@
+local MER = unpack(ElvUI_MerathilisUI)
+
+MER.Changelog[654] = {
+	RELEASE_DATE = "27.02.2025",
+	FIXES = {
+		"[SpecBar]: Fixed an issue which prevent so switch the Specialization- #104",
+	},
+	NEW = {},
+	IMPROVEMENTS = {},
+}