+++ ElvUI_MerathilisUI/Modules/Skins/FrameXML/QuestMapFrame.lua
@@ -45,6 +45,8 @@ function module:QuestMapFrame()
 	end)
 
 	QuestLogPopupDetailFrame.SealMaterialBG:SetAlpha(0)
+
+	_G.QuestMapFrame.CampaignOverview.BG:SetAlpha(0)
 end
 
 module:AddCallback("QuestMapFrame")