+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Changes ###
 
-  * [Objective Tracker]: Fix typo in the options
-  * [Wowhead Links]: Now also works with monthly activities
+  * [Simple AddOn Manager]: Fix an error with the latest code changes
+  * [World Map]: Updated MapData to 11.1
+  * [World Map]: Reenable Event Tracker
+  * [AutoButtons]: Add more openables
+  * [Misc]: Small code cleanups